fix: resolve genre unique constraint in tests and complete movie coverage

+++ cinema/tests/test_movie_api.py
@@ -23,7 +23,6 @@ def sample_movie(**params):
         "duration": 90,
     }
     defaults.update(params)
-
     return Movie.objects.create(**defaults)
 
 
@@ -32,14 +31,12 @@ def sample_genre(**params):
         "name": "Drama",
     }
     defaults.update(params)
-
     return Genre.objects.create(**defaults)
 
 
 def sample_actor(**params):
     defaults = {"first_name": "George", "last_name": "Clooney"}
     defaults.update(params)
-
     return Actor.objects.create(**defaults)
 
 
@@ -47,23 +44,22 @@ def sample_movie_session(**params):
     cinema_hall = CinemaHall.objects.create(
         name="Blue", rows=20, seats_in_row=20
     )
-
     defaults = {
         "show_time": "2022-06-02 14:00:00",
         "movie": None,
         "cinema_hall": cinema_hall,
     }
     defaults.update(params)
-
     return MovieSession.objects.create(**defaults)
 
 
 def image_upload_url(movie_id):
-    """Return URL for recipe image upload"""
+    """Return URL for movie image upload"""
     return reverse("cinema:movie-upload-image", args=[movie_id])
 
 
 def detail_url(movie_id):
+    """Return URL for movie detail"""
     return reverse("cinema:movie-detail", args=[movie_id])
 
 
@@ -84,8 +80,19 @@ class PrivateMovieApiTests(TestCase):
         )
         self.client.force_authenticate(self.user)
 
+    def test_retrieve_movie_detail(self):
+        movie = sample_movie()
+        movie.genres.add(sample_genre(name="Action"))
+        movie.actors.add(sample_actor(first_name="Test", last_name="Actor"))
+
+        url = detail_url(movie.id)
+        res = self.client.get(url)
+
+        serializer = MovieDetailSerializer(movie)
+        self.assertEqual(res.status_code, status.HTTP_200_OK)
+        self.assertEqual(res.data, serializer.data)
+
     def test_list_movies(self):
-        """Test retrieving a list of movies"""
         sample_movie()
         res = self.client.get(MOVIE_URL)
 
@@ -96,7 +103,6 @@ class PrivateMovieApiTests(TestCase):
         self.assertEqual(res.data, serializer.data)
 
     def test_filter_movies_by_title(self):
-        """Test filtering movies by title"""
         sample_movie(title="Avatar")
         sample_movie(title="Batman")
 
@@ -107,7 +113,6 @@ class PrivateMovieApiTests(TestCase):
         self.assertEqual(res.data[0]["title"], "Avatar")
 
     def test_filter_movies_by_genres(self):
-        """Test filtering movies by genres"""
         genre1 = sample_genre(name="Comedy")
         genre2 = sample_genre(name="Sci-Fi")
 
@@ -123,7 +128,6 @@ class PrivateMovieApiTests(TestCase):
         self.assertIn("Comedy", res.data[0]["genres"])
 
     def test_filter_movies_by_actors(self):
-        """Test filtering movies by actors"""
         actor1 = sample_actor(first_name="Brad", last_name="Pitt")
         actor2 = sample_actor(first_name="Tom", last_name="Cruise")
 
@@ -139,7 +143,6 @@ class PrivateMovieApiTests(TestCase):
         self.assertIn("Brad Pitt", res.data[0]["actors"])
 
     def test_create_movie_forbidden(self):
-        """Test that regular users cannot create movies"""
         payload = {
             "title": "Title",
             "description": "Description",
@@ -157,15 +160,15 @@ class MovieImageUploadTests(TestCase):
         )
         self.client.force_authenticate(self.user)
         self.movie = sample_movie()
-        self.genre = sample_genre()
+        self.genre = sample_genre()  # Створює жанр "Drama"
         self.actor = sample_actor()
         self.movie_session = sample_movie_session(movie=self.movie)
 
     def tearDown(self):
-        self.movie.image.delete()
+        if self.movie.image:
+            self.movie.image.delete()
 
     def test_upload_image_to_movie(self):
-        """Test uploading an image to movie"""
         url = image_upload_url(self.movie.id)
         with tempfile.NamedTemporaryFile(suffix=".jpg") as ntf:
             img = Image.new("RGB", (10, 10))
@@ -179,7 +182,6 @@ class MovieImageUploadTests(TestCase):
         self.assertTrue(os.path.exists(self.movie.image.path))
 
     def test_upload_image_bad_request(self):
-        """Test uploading an invalid image"""
         url = image_upload_url(self.movie.id)
         res = self.client.post(url, {"image": "not image"}, format="multipart")
 
@@ -197,8 +199,8 @@ class MovieImageUploadTests(TestCase):
                     "title": "Title",
                     "description": "Description",
                     "duration": 90,
-                    "genres": [1],
-                    "actors": [1],
+                    "genres": [self.genre.id],
+                    "actors": [self.actor.id],
                     "image": ntf,
                 },
                 format="multipart",
@@ -240,3 +242,21 @@ class MovieImageUploadTests(TestCase):
         res = self.client.get(MOVIE_SESSION_URL)
 
         self.assertIn("movie_image", res.data[0].keys())
+
+    def test_create_movie(self):
+        # Використовуємо вже створені self.genre та self.actor
+        payload = {
+            "title": "New Movie",
+            "description": "New Description",
+            "duration": 120,
+            "genres": [self.genre.id],
+            "actors": [self.actor.id],
+        }
+        res = self.client.post(MOVIE_URL, payload)
+
+        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
+        movie = Movie.objects.get(id=res.data["id"])
+
+        self.assertEqual(payload["title"], movie.title)
+        self.assertEqual(movie.genres.count(), 1)
+        self.assertEqual(movie.actors.count(), 1)
